Overwrite ExpSize value in Rounds on tests using small primes
This will prevent panics in the generation phase.
In main_test the value is hardcoded to 6 on tests using p=101.
In benchmarks, the function GenerateRounds will get the p value
from the group, and overwrite the ExpSize only if the number of bits
is smaller than 256. This will effectively only happen in the tests
using p=101.