Fix tests using p=101 in go1.11 and above
With go1.11 the math/big ModInverse function will now return nil
and leave the argument unchanged whenever the numbers involved
are not relatively prime.

Since 101 is not a strong prime, there is a high chance that the
FindSmallCoprimeInverse function will hit the condition above, meaning
that the round.Z value will be unchanged.

This way, it is possible to compare that value to its default of
MaxInt, in order to know that the generation was not successful.
Modified the tests using p=101 to loop the generation procedure
until valid Z keys are generated.