Resurrect cmd and benchmark glue code, disabling or commenting out anything we don't use/need any more.
NOTE: I left things we want to turn back on, but don't have infrastructure in place to do so, as commented out code. This includes specific command line functions (and their deps) as well as a number of tests. As we add these back in, we should delete the commented out stuff as we add these features back in, so it is tractable what we are still missing from the rewrite.