Fix precanned permutations for edge case and fix realtime decrypt
Precanned permutations would case access out of bounds if the input sizes
were different and larger than the batch size in a specific way. It should
no longer be an issue.

Realtime decrypt was putting output in a variable that it shouldn't,
which caused results to be incorrect later. Now it doesn't do that.